requirements.txt
absl-py==1.2.0
addict==2.4.0
anyio==3.6.1
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
chart-studio==1.1.0
colorama==0.4.5
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.7.0
entrypoints==0.4
etils==0.8.0
evo==1.19.0
executing==1.0.0
faiss==1.7.2
fastjsonschema==2.16.1
fonttools==4.37.1
gtsam==4.1.1
h5py==3.7.0
idna==3.3
imageio==2.21.1
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jax==0.3.17
jaxlib==0.3.15
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonschema==4.14.0
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyter_packaging==0.12.3
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
jupyterlab_server==2.15.1
kiwisolver==1.4.4
lxml==4.9.1
lz4==4.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
natsort==8.2.0
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.6
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.2
numpy-quaternion @ file:///home/conda/feedstock_root/build_artifacts/quaternion_1651115094299/work
open3d==0.15.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
plotly==5.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydegensac==0.1.2
Pygments==2.13.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
Quaternion==3.5.2.post4
requests==2.28.1
retrying==1.3.3
rosbags==0.9.12
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-build==0.15.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
Send2Trash==1.8.0
Shapely==1.8.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tabulate==0.8.10
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.1.1
tomlkit==0.11.4
torch==1.7.0
torchvision==0.8.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
widgetsnbextension==4.0.2
zipp==3.8.1
zstandard==0.18.0
